feat: add SunCard component to display sun data on the dashboard
